Extract plugin management into a self-contained plugin_management plugin
Move all plugin-management code (PluginGroupManager, manifest/paths/installer,
the manage/uninstall dialog models, and the Install/Uninstall/Manage actions)
out of microdrop_application into a new top-level plugin_management package.
PluginManagementPlugin offers the PluginGroupManager service and contributes
the three Tools actions via TASK_EXTENSIONS (manual_controls pattern), and
runs the launch-restore from its application_initialized hook. The Tools menu
in MicrodropTask is now an empty anchor; microdrop_application no longer
references any plugin-management code (one-way dependency). Added to
FRONTEND_PLUGINS. No behavior change.
