--- microdrop_application/plugins/__init__.py
@@ -1,1 +1,0 @@
-"""Runtime plugin install + manifest-driven group discovery."""
