+++ plugin_management/__init__.py
@@ -1,0 +1,2 @@
+"""Runtime plugin management: discover, install, uninstall, and hot
+load/unload plugin groups, contributed to the app as a self-contained plugin."""
