+++ plugin_management/consts.py
@@ -1,0 +1,3 @@
+# This module's package.
+PKG = '.'.join(__name__.split('.')[:-1])
+PKG_name = PKG.title().replace("_", " ")
