+++ plugin_management/manage_dialog.py
