+++ plugin_management/manifest.py
