+++ plugin_management/uninstall_dialog.py
